Remove reference url json generator, and prep for moving schematron assertion reference URL into .sch -> .json summary generator.
