+++ src/web/src/shared/use-cases/schematron-summary.test.ts
@@ -1,0 +1,5 @@
+import { describe, it } from 'vitest';
+
+describe('schematron summary', () => {
+  it('works', () => {});
+});
